Implement service test

+++ src/main/java/com/doodle/poll/domain/InitiatorRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InitiatorRepo extends CrudRepository<Initiator, Long>{
+public interface InitiatorRepo extends CrudRepository<Initiator, Integer>{
 
 	List<Initiator> findByEmail(String user);
