+++ src/main/java/com/doodle/poll/domain/ParticipantRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParticipantRepo extends CrudRepository<Participant, Long>{
+public interface ParticipantRepo extends CrudRepository<Participant, Integer>{
 
 }
